# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: API documentation for 3X-UI Panel (Master-Slave Cluster)
  title: 3X-UI Cluster API
  version: "1.0"
paths:
...
  /panel/api/server/cpuHistory/{bucket}:
    get:
      description: Returns aggregated CPU usage history for chart display
      parameters:
      - description: Time bucket in seconds (2,30,60,120,180,300)
        in: path
        name: bucket
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get CPU history
      tags:
      - Server
...
  /panel/api/server/getNewEchCert:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Generates a new ECH (Encrypted Client Hello) certificate
      parameters:
      - description: Server Name Indication
        in: formData
        name: sni
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Generate ECH certificate
      tags:
      - Server
...
  /panel/api/server/importDB:
    post:
      consumes:
      - multipart/form-data
      description: Imports a database file and restarts the Xray service
      parameters:
      - description: Database file
        in: formData
        name: db
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Import database
      tags:
      - Server
  /panel/api/server/installXray/{version}:
    post:
      description: Installs or updates Xray to the specified version
      parameters:
      - description: Xray version to install
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Install Xray version
      tags:
      - Server
  /panel/api/server/logs/{count}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Retrieves panel logs with filtering by count, level, and syslog
      parameters:
      - description: Number of log lines
        in: path
        name: count
        required: true
        type: string
      - description: Log level filter
        in: formData
        name: level
        type: string
      - description: Syslog filter
        in: formData
        name: syslog
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get application logs
      tags:
      - Server
  /panel/api/server/restartSlaveXray/{slaveId}:
    post:
      description: Restarts the Xray service on a specific slave node
      parameters:
      - description: Slave ID
        in: path
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Restart slave Xray
      tags:
      - Server
...
  /panel/api/server/updateGeofile/{fileName}:
    post:
      description: Updates the specified geo file (geoip.dat/geosite.dat) for Xray
      parameters:
      - description: Geo file name
        in: path
        name: fileName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Update geo file
      tags:
      - Server
  /panel/api/server/xraylogs/{count}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Retrieves Xray access logs with filtering by traffic type
      parameters:
      - description: Number of log lines
        in: path
        name: count
        required: true
        type: string
      - description: Text filter
        in: formData
        name: filter
        type: string
      - description: Show direct traffic
        in: formData
        name: showDirect
        type: string
      - description: Show blocked traffic
        in: formData
        name: showBlocked
        type: string
      - description: Show proxy traffic
        in: formData
        name: showProxy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get Xray logs
      tags:
      - Server
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: 3x-ui
    type: apiKey
swagger: "2.0"